+++ trikot-kword/kword-plugin/src/main/kotlin/com.mirego.kword/KWordPlugin.kt
@@ -4,16 +4,21 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 
 class KWordPlugin : Plugin<Project> {
-    companion object {
+    private companion object {
         private const val TASKS_GROUP = "KWord"
     }
 
     override fun apply(project: Project) {
-        val extension = project.extensions.create("kword", KWordExtension::class.java, project)
+        val extension = project.extensions.create("kword", KWordExtension::class.java)
 
         project.tasks.create("kwordGenerateEnum", KWordEnumGenerate::class.java) {
             it.group = TASKS_GROUP
             it.description = "Generate keys enum based on json translation file."
+            it.enumClassName.set(extension.enumClassName)
+            it.translationFiles.setFrom(
+                extension.translationFiles + extension.translationFile
+            )
+            it.generatedDir.set(extension.generatedDir)
         }
     }
 }