+++ trikot-kword/kword-plugin/src/main/kotlin/com.mirego.kword/KwordExtension.kt
@@ -1,38 +1,16 @@
 package com.mirego.kword
 
-import org.gradle.api.Project
-import org.gradle.api.file.Directory
-import org.gradle.api.tasks.Input
-import java.io.File
+import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
 
-open class KWordExtension(project: Project) {
-    @Input
-    var translationFile: File? = null
+abstract class KWordExtension {
+    abstract val translationFile: RegularFileProperty
 
-    @Input
-    var translationFiles: List<File> = emptyList()
+    abstract val translationFiles: ConfigurableFileCollection
 
-    @Input
-    var enumClassName: String? = null
+    abstract val enumClassName: Property<String>
 
-    @Input
-    var generatedDir: File? = null
-
-    private val projectDirectory: Directory = project.layout.projectDirectory
-
-    fun translationFile(translationFile: Any) {
-        this.translationFile = projectDirectory.files(translationFile).singleFile
-    }
-
-    fun translationFiles(vararg translationFiles: Any) {
-        this.translationFiles = projectDirectory.files(*translationFiles).toList()
-    }
-
-    fun enumClassName(enumClassName: String) {
-        this.enumClassName = enumClassName
-    }
-
-    fun generatedDir(generatedDir: Any) {
-        this.generatedDir = projectDirectory.files(generatedDir).singleFile
-    }
+    abstract val generatedDir: DirectoryProperty
 }